fix: avoid start rendering thankyou page if there is a custom one
