+++ .circleci/config.yml
@@ -7,7 +7,7 @@ orbs:
 jobs:
   test: # this can be any name you choose
     docker:
-      - image: mcr.microsoft.com/playwright:v1.52.0-noble
+      - image: mcr.microsoft.com/playwright:v1.53.0-noble
     resource_class: medium+
     parallelism: 10
 
@@ -50,7 +50,7 @@ jobs:
           no_output_timeout: 15m
           command: |
             mkdir test-results
-            pnpm run serve &
+            pnpm run serve --no-request-logging &
             TESTFILES=$(circleci tests glob "specs/e2e/**/*.spec.ts")
             echo $TESTFILES | circleci tests run --command="xargs pnpm playwright test --config=playwright.config.ci.ts --reporter=list" --verbose --split-by=timings
       - store_test_results:
